+++ protected/humhub/docs/CHANGELOG.md
@@ -33,6 +33,10 @@ HumHub Change Log
 - Enh: Added upload file event (humhub:file:created)
 - Enh: Added custom file handler positions
 - Enh: Moved UploadAction::getFileResponse method to FileHelper::getFileInfos
+- Enh: Added JS context menu to ui.additions module
+- Enh: Enhanced ContentContainer Module enable/disable
+- Enh: Added client.reload for pjax and non pjax page reloads
+- Enh: Added ContentContainerAsset to AppAsset
 
 1.2.0-beta.1 (February 08, 2017)
 --------------------------------